Split error_handling.hpp into exceptions.hpp and safe_call.hpp.
error_handling.cpp was simply renamed to exceptions.cpp, since it contained
neither implemented nor used the content of safe_call.hpp.

These changes were made based on a dichotomy in the purposes of users of the old
file, as well as a desire to limit the scope of any JNI dependencies.